Create SEARCH NUMBER
Write the pseudo-code for an algorithm that performs a sequential search to find a match for a 1.
user-specified value in the following list:0,3,7,2,1,12,50.
Represent this list as L0..n−1. n is the length of the list (7). If the user specifies a number that is
in the list, the algorithm should output the offset of the match. For example, if the user inputs
0, the output is 0. If the user inputs 7,the output is 2. If the user inputs a number that does not
appear in the list, the output should be n, which in this case is 7.


Implement binary search in assembly language so that a user can input a number and the 4.
algorithm will respond as with the sequential search algorithm. You will need to make use of
the non-standard LOAD 0 instruction that uses the R register as a pointer. Store the following
list in consecutive .DATA blocks: 0,1,2,3,7,10,12,38,50. Submit the assembly code and a copy
and paste of the algorithm output.